requirements-core.txt
textX==4.2.0
Jinja2==3.1.5
river==0.22.0
paho-mqtt==2.1.0
redis==6.1.0
numpy==2.1.3
pandas==2.2.3
scikit-learn==1.6.1
scipy==1.15.3
tensorflow==2.19.0
keras==3.9.2

